requirements.txt
attrs==22.2.0
autopep8==2.0.2
cachelib==0.9.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.2
Deprecated==1.2.13
dnspython==2.3.0
exceptiongroup==1.1.1
Flask==2.2.3
flask-expects-json==1.7.0
Flask-Limiter==3.3.0
gunicorn==20.1.0
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
limits==3.3.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mongomock==4.1.2
ordered-set==4.1.0
packaging==23.0
platformdirs==3.1.1
pluggy==1.0.0
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pymongo==4.3.3
pyrsistent==0.19.3
pytest==7.2.2
pytest-cov==4.0.0
python-dotenv==1.0.0
requests==2.28.2
requests-cache==1.0.0
requests-mock==1.10.0
rich==13.3.2
sentinels==1.0.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
url-normalize==1.4.3
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
